requirements.txt
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3
tensorflow==2.6.0
streamlit==0.88.0
Pillow==8.3.2